fix: all issues from testing fixed, changed phpunit db from sqlite to pgsql

+++ database/migrations/2026_03_02_191536_create_themes_table.php
@@ -11,12 +11,6 @@ return new class extends Migration
      */
     public function up(): void
     {
-        DB::statement('CREATE EXTENSION IF NOT EXISTS pg_trgm;');
-
-        DB::statement('CREATE INDEX themes_title_trgm_idx ON themes USING GIN (title gin_trgm_ops);');
-
-        DB::statement('CREATE INDEX themes_description_trgm_idx ON themes USING GIN (description gin_trgm_ops);');
-
         Schema::create('themes', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->nullable()->constrained('users')->nullOnDelete();
@@ -26,6 +20,14 @@ return new class extends Migration
             $table->json('images')->nullable();
             $table->timestamps();
         });
+
+        if (DB::connection()->getDriverName() === 'pgsql') {
+            DB::statement('CREATE EXTENSION IF NOT EXISTS pg_trgm;');
+
+            DB::statement('CREATE INDEX themes_title_trgm_idx ON themes USING GIN (title gin_trgm_ops);');
+
+            DB::statement('CREATE INDEX themes_description_trgm_idx ON themes USING GIN (description gin_trgm_ops);');
+        }
     }
 
     /**
@@ -33,8 +35,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        DB::statement('DROP INDEX IF EXISTS themes_title_trgm_idx');
-        DB::statement('DROP INDEX IF EXISTS themes_description_trgm_idx');
+        if (DB::connection()->getDriverName() === 'pgsql') {
+            DB::statement('DROP INDEX IF EXISTS themes_title_trgm_idx');
+            DB::statement('DROP INDEX IF EXISTS themes_description_trgm_idx');
+        }
+
         Schema::dropIfExists('themes');
     }
 };
